1. Cust 模型，增加门店签到状态 2. 签到增加两点间距离判断 3. 增加店名搜索门店列表 4.完善上下班打卡，门店签到流程

+++ database/migrations/2015_06_19_100149_create_attn_table.php
@@ -56,10 +56,10 @@ class CreateAttnTable extends Migration
             ])->default(Attn::STATUS_NG);
 
             // 里程数
-            $table->tinyInteger('mileage')->default(0);
+            $table->decimal('mileage',8,2)->default(0);
 
             // 巡店数
-            $table->tinyInteger('visited_custs')->default(0);
+            $table->integer('visited_custs')->default(0);
 
             $table->timestamps();
         });
